first issue of my bashrc-file with its associates
Instruction: Just download the file bashrc.tar.xz and extract it in your home folder.
Suggestions are always welcome.

This bashrc consists of several files. On the one hand the actual bashrc and on the other hand a few files that are assigned to certain settings. In principle, the contents of these files could also be written directly to the bashrc, but it is a bit clearer this way. These files can be found in the ~/shellcfg/ directory.

The ~/shellcfg/what_shell file is actually nonsense because other shells have different file names and the output could have been a simple query (like "$SHELL"). Furthermore the version query does only work in BASH this way. However, it doesn't bother me either, so I left it there.

If you want to modify the bashrc and its associated files for other shells, feel free to do so. Just fork this repo and send me some love when you do so.