Using check for `vivid` and use it, when installed
this update changes how we configure LS_COLORS. Now we check for vivid and if it is installed we use it with the theme `molokai`. If `vivid` cannot be found, the colors for `ls` are defined in the conventional way and exported into the LS_COLORS variable.
Furthermore we define variables at first and export them afterwards. This ensures that we are POSIX compatible  as much as possible.